refactoring

+++ src/main/java/com/opensource/projectu/service/ProjectService.java
@@ -5,6 +5,8 @@ import com.opensource.projectu.openapi.model.Project;
 import com.opensource.projectu.repository.ProjectRepository;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
 import java.util.List;
 import java.util.UUID;
 
@@ -24,33 +26,43 @@ public class ProjectService {
     }
 
     public Project createProject(Project project) {
-        project.setId(UUID.randomUUID().toString());
+        project.setId(generateUniqueId());
         return projectRepository.save(project);
     }
 
-    public Project updateProject(String id, Project newProject) {
+    private String generateUniqueId() {
+        var id = UUID.randomUUID().toString();
+        if(!projectRepository.existsById(id)) {
+            return id;
+        }
+        return generateUniqueId();
+    }
+
+    @Transactional
+    public Project updateProject(String id, Project updatedProject) {
         var currentProject = projectRepository.findById(id)
                 .orElseThrow(() -> new NotFoundException("Project with id "+ id +" not found."));
 
-        var updatedProject = copyProjectProperties(newProject, currentProject);
-        return projectRepository.save(updatedProject);
-    }
+        // TODO: validate each property
+        currentProject
+                .title(updatedProject.getTitle())
+                .description(updatedProject.getDescription())
+                .tasks(updatedProject.getTasks())
+                .state(updatedProject.getState())
+                .complexity(updatedProject.getComplexity())
+                .estimatedDurationInHours(updatedProject.getEstimatedDurationInHours())
+                .expectedResult(updatedProject.getExpectedResult())
+                .actualResult(updatedProject.getActualResult())
+                .startedAt(updatedProject.getStartedAt())
+                .finishedAt(updatedProject.getFinishedAt());
 
-    private Project copyProjectProperties(Project newProject, Project currentProject) {
-        currentProject.setTitle(newProject.getTitle());
-        currentProject.setDescription(newProject.getDescription());
-        currentProject.setTasks(newProject.getTasks());
-        currentProject.setState(newProject.getState());
-        currentProject.setComplexity(newProject.getComplexity());
-        currentProject.setEstimatedDurationInHours(newProject.getEstimatedDurationInHours());
-        currentProject.setExpectedResult(newProject.getExpectedResult());
-        currentProject.setActualResult(newProject.getActualResult());
-        currentProject.setStartedAt(newProject.getStartedAt());
-        currentProject.setFinishedAt(newProject.getFinishedAt());
-        return currentProject;
+        return projectRepository.save(currentProject);
     }
 
     public void deleteProject(String id) {
+        if(!projectRepository.existsById(id)) {
+            throw new NotFoundException("Project with id "+ id +" not found.");
+        }
         projectRepository.deleteById(id);
     }
 }
